requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysql-connector-python
python-dotenv